Move DartSchemaType and subclasses to their own library
