+++ discoveryapis_generator/lib/src/dart_resources.dart
@@ -10,6 +10,7 @@ import 'dart:collection';
 
 import 'dart_api_library.dart';
 import 'dart_comments.dart';
+import 'dart_schema_types.dart';
 import 'dart_schemas.dart';
 import 'generated_googleapis/discovery/v1.dart';
 import 'namer.dart';
