+++ discoveryapis_generator/test/dart_resources_test.dart
@@ -4,6 +4,7 @@
 
 import 'package:discoveryapis_generator/src/dart_api_library.dart';
 import 'package:discoveryapis_generator/src/dart_resources.dart';
+import 'package:discoveryapis_generator/src/dart_schema_types.dart';
 import 'package:discoveryapis_generator/src/dart_schemas.dart';
 import 'package:discoveryapis_generator/src'
     '/generated_googleapis/discovery/v1.dart';
